GridapMakie

The documentation of this package is under construction. For further information read the README.md file.

GridapMakie.MeshField

GridapMakie.meshfield

GridapMakie.meshfield!

GridapMakie.MeshField — Type

MeshField is the plot type associated with plotting function meshfield. Check the docstring for meshfield for further information.

GridapMakie.meshfield — Function

No docstring defined.

Plot type

The plot type alias for the meshfield function is MeshField.

Attributes

alpha = 1.0 — The alpha value of the colormap or color attribute. Multiple alphas like in plot(alpha=0.2, color=(:red, 0.5), will get multiplied.

backlight = 0.0 — Sets a weight for secondary light calculation with inverted normals.

clip_planes = @inherit clip_planes automatic — Clip planes offer a way to do clipping in 3D space. You can set a Vector of up to 8 Plane3f planes here, behind which plots will be clipped (i.e. become invisible). By default clip planes are inherited from the parent plot or scene. You can remove parent clip_planes by passing Plane3f[].

color = :pink — No docs available.

colormap = :bluesreds — Sets the colormap that is sampled for numeric colors. PlotUtils.cgrad(...), Makie.Reverse(any_colormap) can be used as well, or any symbol from ColorBrewer or PlotUtils. To see all available color gradients, you can call Makie.available_gradients().

colorrange = automatic — The values representing the start and end points of colormap.

colorscale = identity — The color transform function. Can be any function, but only works well together with Colorbar for identity, log, log2, log10, sqrt, logit, Makie.pseudolog10 and Makie.Symlog10.

cycle = nothing — No docs available.

depth_shift = 0.0 — Adjusts the depth value of a plot after all other transformations, i.e. in clip space, where -1 <= depth <= 1. This only applies to GLMakie and WGLMakie and can be used to adjust render order (like a tunable overdraw).

depthsorting = false — No docs available.

diffuse = 1.0 — Sets how strongly the red, green and blue channel react to diffuse (scattered) light.

distancefield = nothing — No docs available.

font = "default" — No docs available.

fxaa = false — Adjusts whether the plot is rendered with fxaa (anti-aliasing, GLMakie only).

glowcolor = (:black, 0.0) — No docs available.

glowwidth = 0.0 — No docs available.

highclip = automatic — The color for any value above the colorrange.

inspectable = @inherit inspectable — Sets whether this plot should be seen by DataInspector. The default depends on the theme of the parent scene.

inspector_clear = automatic — Sets a callback function (inspector, plot) -> ... for cleaning up custom indicators in DataInspector.

inspector_hover = automatic — Sets a callback function (inspector, plot, index) -> ... which replaces the default show_data methods.

inspector_label = automatic — Sets a callback function (plot, index, position) -> string which replaces the default label generated by DataInspector.

interpolate = true — No docs available.

joinstyle = @inherit joinstyle — No docs available.

linecap = @inherit linecap — No docs available.

linestyle = nothing — No docs available.

linewidth = @inherit linewidth — No docs available.

lowclip = automatic — The color for any value below the colorrange.

marker = @inherit marker — No docs available.

marker_offset = Makie.Vec3f(0) — No docs available.

markersize = @inherit markersize — No docs available.

markerspace = :pixel — No docs available.

matcap = nothing — No docs available.

material = nothing — RPRMakie only attribute to set complex RadeonProRender materials. Warning, how to set an RPR material may change and other backends will ignore this attribute

miter_limit = @inherit miter_limit — No docs available.

model = automatic — Sets a model matrix for the plot. This overrides adjustments made with translate!, rotate! and scale!.

nan_color = :transparent — The color for NaN values.

overdraw = false — Controls if the plot will draw over other plots. This specifically means ignoring depth checks in GL backends

rotation = Makie.Billboard() — No docs available.

shading = Makie.NoShading — Controls if the plot object is shaded by the parent scenes lights or not. The lighting algorithm used is controlled by the scenes shading attribute.

shininess = 32.0 — Sets how sharp the reflection is.

space = :data — Sets the transformation space for box encompassing the plot. See Makie.spaces() for possible inputs.

specular = 0.2 — Sets how strongly the object reflects light in the red, green and blue channels.

ssao = false — Adjusts whether the plot is rendered with ssao (screen space ambient occlusion). Note that this only makes sense in 3D plots and is only applicable with fxaa = true.

strokecolor = @inherit markerstrokecolor — No docs available.

strokewidth = @inherit markerstrokewidth — No docs available.

transformation = :automatic — No docs available.

transparency = false — Adjusts how the plot deals with transparency. In GLMakie transparency = true results in using Order Independent Transparency.

uv_offset_width = (0.0, 0.0, 0.0, 0.0) — No docs available.

uv_transform = Makie.automatic — No docs available.

visible = true — Controls whether the plot will be rendered or not.

GridapMakie.meshfield! — Function

meshfield! is the mutating variant of plotting function meshfield. Check the docstring for meshfield for further information.